Feature: Update nib with better formatting to display Observation information add method in SearchViewModel to load an icon add extension method to UIImage to load a remote image update StationDetailViewRepresentable to display information from an observation

+++ HereApp/HereApp/Core/Models/WeatherResponse.swift
@@ -72,6 +72,6 @@ struct WeatherPlace: Decodable {
 struct Observation: Decodable {
     
     let description, iconName, iconLink: String
-    let temperature: Int
+    let temperature: Double
     
 }
